requirements.txt
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
pyyaml==6.0
jupyterlab==4.3.1
ipykernel==6.29.5
torch==2.3.0
clickhouse-driver==0.2.7